merge

See merge request projets/2023_gp_2027_gp_2027_projet_moteur-undefined!15

+++ Undefined/source/src/interface/FPSgraph.cpp
@@ -9,7 +9,7 @@
 void FPSGraph::Init()
 {
     mUpdateTime = 0.1f;
-    mLastTotalTime = glfwGetTime();
+    mLastTotalTime = (float)glfwGetTime();
     mMaxFPS = 200.0f;
     mArraySize = 50;
 }
@@ -18,7 +18,7 @@ void FPSGraph::ShowWindow()
 {
     ImGui::Begin("FPS Graph");
  
-    // Every ImGUI slider 
+    //Every ImGUI slider 
     ImGui::SliderFloat("Time between updates", &mUpdateTime, 0.0f, 1.0f, "%.2f");
     ImGui::SliderFloat("Max FPS", &mMaxFPS, 0.0f, 300, "%.2f");
     if (ImGui::SliderInt("Array Size", &mArraySize, 2, 100))
@@ -26,29 +26,31 @@ void FPSGraph::ShowWindow()
         mFrameRateArray.resize(mArraySize);
     }
 
-    if (Utils::OnInterval(glfwGetTime(), mLastTotalTime, mUpdateTime))
+    if (Utils::OnInterval((float)glfwGetTime(), mLastTotalTime, mUpdateTime))
     {
         float lastFPS = 1.0f / ImGui::GetIO().DeltaTime;
 
+        //If the array is not full we pushback values instead of going to an index to modify the value so the array fills up and display a correct fps average and not an average with multiple zero's in the array
         if (mFrameRateArray.size() < mArraySize)
             mFrameRateArray.push_back(lastFPS);
         
-        // We check if the ArrayIndex is in the FrameRateArray size so there is no vector out of range because of call stack when we change ArraySize
+        //We check if the ArrayIndex is in the FrameRateArray size so there is no vector out of range because of call stack when we change ArraySize
         if(mArrayIndex <= mFrameRateArray.size())
             mFrameRateArray[mArrayIndex] = lastFPS;
 
         mArrayIndex = (mArrayIndex + 1) % mFrameRateArray.size();
-        mLastTotalTime = glfwGetTime();
+        mLastTotalTime = (float)glfwGetTime();
     }
 
     std::string string = "average : " + std::to_string((int)Utils::Average(mFrameRateArray));
 
+    //Cap the min fps to : current fps + 5
     if (mMaxFPS < Utils::Average(mFrameRateArray))
     {
         mMaxFPS = Utils::Average(mFrameRateArray) + 5;
     }
 
-    // Draw FPS Graph
+    //Draw FPS Graph
     ImGui::PlotLines("", mFrameRateArray.data(), (int)mFrameRateArray.size(), 0, string.c_str(), 0, mMaxFPS, ImGui::GetContentRegionAvail());
 
     ImGui::End();
